Added Files via Upload (SMA)
Uploaded a program I have been working on SMA.py. The purpose of this program is to determine the Simple Moving Average of a stock price and visualise it in a Graph.

The program collects the stock price data from Yahoo Finance via the yfinance python module. The data is stored in a dataframe and then the program appends the dataframe provided by yfinance to include the Next Days Closing Price, The Daily Price Change and the Daily Returns. Finally the program determines the Moving Day Averages for the days set. (Ideally One Day being a Fast Signal and the other being a Slow Signal)

The program saves this data into a CSV File, and also generates a graph visualising the two Moving Day Averages and the Stocks Closing Price over a selected period of time. I have uploaded examples for stocks Amazon, Apple, Google and Facebook.